LPD-19899 Add upgrade process to transform the 'applicationName' and 'scopeAliases' fields to lowercase when they are related to Object's bundle.
This is necessary as we updated the JAX-RS application name for objects. As now for any object with the same lowercase name, the JAX-RS application name must be the same, so we unified all the JAX-RS application names based on the lowercase Object name.

For example:
- C_MyTest (in a company #1)
- C_mytest (in a company #2)

both objects are now managed by the same JAX-RS application name with path /c/mytests, so the JAX-RS application name is 'mytest'.