LPD-63423 Rename method and variable so it reflects the logic being applied.
In headless-delivery FragmentMappedValueUtil the method is called "isSaveFragmentMappedValue" because it takes into account the user input "saveMapping" parameter to determine whether mapping should be saved or not.

Within this class we always want to save the mapping, so what the method is determining is whether the value configured corresponds to a mapping or not.